# backend/src/config/swagger.yaml
openapi: 3.0.0
info:
  title: ReachstreamAPI
  description: Real-Time Social Media Data API - Extract public data from TikTok, Instagram, YouTube, Twitter, Facebook, LinkedIn, and Reddit
  version: 1.0.0
  contact:
    name: ReachstreamAPI Support
    email: support@reachstreamapi.com
    url: https://reachstream.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.reachstream.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authentication
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        metadata:
          $ref: '#/components/schemas/Metadata'
    Metadata:
      type: object
      properties:
        response_time_ms:
          type: integer
          example: 2341
        proxy_used:
          type: boolean
          example: true
        timestamp:
          type: string
          format: date-time
          example: "2025-10-31T17:00:00Z"
security:
  - ApiKeyAuth: []
paths:
...
  /api/scrape/tiktok/profile:
    get:
      tags:
        - TikTok
      summary: Get TikTok profile data
      description: Retrieve complete TikTok profile information including follower count, bio, and statistics
      parameters:
        - name: username
          in: query
          required: true
          description: TikTok username (without @)
          schema:
            type: string
            example: charlidamelio
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user_id:
                        type: string
                      username:
                        type: string
                      nickname:
                        type: string
                      avatar_url:
                        type: string
                      signature:
                        type: string
                      verified:
                        type: boolean
                      private_account:
                        type: boolean
                      follower_count:
                        type: integer
                      following_count:
                        type: integer
                      video_count:
                        type: integer
                      heart_count:
                        type: integer
                      profile_url:
                        type: string
                      scraped_at:
                        type: string
                        format: date-time
                  metadata:
                    $ref: '#/components/schemas/Metadata'
        '400':
          description: Bad request - missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...